+++ src/common/common.ts
@@ -29,6 +29,17 @@ export type JsonLd = {
   [property: string]: unknown;
 };
 
+type Proof = {
+  type: string;
+  /**
+   * ISO-8601 formatted date
+   */
+  created: string;
+  verificationMethod: string;
+  proofPurpose: string;
+  proofValue: string;
+};
+
 /**
  * A Verifiable Credential JSON-LD document, as specified by the W3C VC HTTP API.
  */
@@ -50,16 +61,15 @@ export type VerifiableCredential = JsonLd & {
      */
     [property: string]: unknown;
   };
-  proof: {
-    type: string;
-    /**
-     * ISO-8601 formatted date
-     */
-    created: string;
-    verificationMethod: string;
-    proofPurpose: string;
-    proofValue: string;
-  };
+  proof: Proof;
+};
+
+export type VerifiablePresentation = JsonLd & {
+  id?: string;
+  type: string[];
+  verifiableCredential?: VerifiableCredential[];
+  holder?: string;
+  proof?: Proof;
 };
 
 /**
@@ -107,6 +117,43 @@ export function isVerifiableCredential(
     dataIsVc &&
     typeof (data as VerifiableCredential).proof.verificationMethod === "string";
   return dataIsVc;
+}
+
+function isUrl(url: string): boolean {
+  try {
+    // If url is not URL-shaped, this will throw.
+    // eslint-disable-next-line no-new
+    new URL(url);
+    return true;
+  } catch (_e) {
+    return false;
+  }
+}
+
+export function isVerifiablePresentation(
+  vp: unknown | VerifiablePresentation
+): vp is VerifiablePresentation {
+  let inputIsVp = true;
+  inputIsVp = inputIsVp && Array.isArray((vp as VerifiablePresentation).type);
+  if ((vp as VerifiablePresentation).verifiableCredential !== undefined) {
+    inputIsVp =
+      inputIsVp &&
+      Array.isArray((vp as VerifiablePresentation).verifiableCredential);
+    inputIsVp =
+      inputIsVp &&
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      (vp as VerifiablePresentation).verifiableCredential!.every(
+        isVerifiableCredential
+      );
+  }
+  if ((vp as VerifiablePresentation).holder !== undefined) {
+    inputIsVp =
+      inputIsVp && typeof (vp as VerifiablePresentation).holder === "string";
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    inputIsVp = inputIsVp && isUrl((vp as VerifiablePresentation).holder!);
+  }
+  // TODO: No type checking is currently implemented for the proof.
+  return inputIsVp;
 }
 
 export function concatenateContexts(...contexts: unknown[]): unknown {